Move bigIntToNumber to utils
